Remove PelletServerTest
It's only used once by PelletClientTest so it might as well exist there.
