+++ client/src/test/java/com/complexible/pellet/client/PelletClientTest.java
@@ -1,17 +1,24 @@
 package com.complexible.pellet.client;
 
-import com.clarkparsia.pellet.server.PelletServerTest;
+import com.clarkparsia.pellet.server.PelletServer;
+import com.clarkparsia.pellet.server.PelletServerModule;
+import com.clarkparsia.pellet.server.TestModule;
+import com.clarkparsia.pellet.server.protege.ProtegeServerTest;
+import com.google.inject.Guice;
+import com.google.inject.util.Modules;
 import edu.stanford.protege.metaproject.ConfigurationManager;
 import edu.stanford.protege.metaproject.api.PlainPassword;
 import edu.stanford.protege.metaproject.api.PolicyFactory;
 import edu.stanford.protege.metaproject.api.UserId;
+import org.junit.After;
 import org.junit.Before;
 import org.protege.editor.owl.client.LocalHttpClient;
 
 /**
  * @author Edgar Rodriguez-Diaz
  */
-public abstract class PelletClientTest extends PelletServerTest {
+public abstract class PelletClientTest extends ProtegeServerTest {
+	protected static PelletServer pelletServer;
 	protected PelletServiceProvider serviceProvider = new PelletServiceProvider(PelletService.DEFAULT_LOCAL_ENDPOINT, 0, 0, 0); // disable all timeouts for tests
 
 	protected LocalHttpClient mClient;
@@ -26,5 +33,24 @@ public abstract class PelletClientTest extends PelletServerTest {
 		UserId managerId = f.getUserId("bob");
 		PlainPassword managerPassword = f.getPlainPassword("bob");
 		managerClient = new LocalHttpClient(managerId.get(), managerPassword.getPassword(), "http://localhost:8081");
+	}
+
+	public void startPelletServer(String... ontologies) throws Exception {
+		pelletServer = new PelletServer(Guice.createInjector(Modules.override(new PelletServerModule())
+		                                                            .with(new TestModule(ontologies))));
+		pelletServer.start();
+	}
+
+	public void stopPelletServer() {
+		if (pelletServer != null) {
+			pelletServer.stop();
+			pelletServer = null;
+		}
+	}
+
+	@After
+	public void after() throws Exception {
+		stopPelletServer();
+		super.after();
 	}
 }
